refactor(tui): refactor task inputs into single component
